Multivariate Cholesky Block Decomposition

mvchol returns modified Cholesky block decomposition (MCBD) of the sample covariance matrix for multivariate longitudinal data with j outcomes.

mvchol(S,t,j)

Arguments

S
Sample covariance matrix of the multivariate longitudinal data.

t
positive integer indicating the number of repeated measurements.

j
positive integer representing the number of outcomes.

Value

Elements from MCBD of the covariance matrix:

T.mat is the regression coefificients matrix \(T\) matrix in the MCBD.

Phit is \(\Phi_{t}\) matrix with regression coefficients stacked as \(\phi_{t1}, \phi_{t2}, ..., \phi_{tk}\).

bigPhi is \(\Phi\) matrix with regression coefficients stacked as \(\Phi_{2}, \Phi_{3}, ..., \Phi_{T}\).

D is the innovation variance matrix with block diagonal elements \(D_1, D_2, \ldots , D_t\).

Dt are the j*j innovation variance matrices.

References

Kohli, P. Garcia, T. and Pourahmadi, M. 2016 Modeling the Cholesky Factors of Covariance Matrices of Multivariate Longitudinal Data, Journal of Multivariate Analysis, 145, 87-100.

Examples

data(Tcells)
S <- cov(Tcells)
t <- 10
j <- 4
MCBD <- mvchol(S,t,j)
names(MCBD)

#> [1] "T.mat"  "Phit"   "bigphi" "D"      "Dt"